pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "nvidia-physicsnemo-cfd"
authors = [
  { name="NVIDIA PhysicsNeMo Team"},
]
description = "A library to use PhysicsNeMo models for engineering and CFD workflows"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "numpy>=1.22.4",
    "pyvista>=0.45.2",
    "tqdm>=4.67.1",
    "numba>=0.60.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "matplotlib>=3.10.3",
    "cupy-cuda12x>=13.4.1",
    "cuml-cu12>=25.4.0",
    "warp-lang>=1.7.1",
    "trimesh>=4.6.9",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dynamic = ["version"]

[tool.setuptools.dynamic]
version = {attr = "physicsnemo.cfd.__version__"}

[tool.setuptools.packages.find]
include = ["physicsnemo.*"]
